Float pos + does not plot out of scope pixels

+++ components/DAC_XY_Plotter/DAC_XY_Plotter.c
@@ -41,7 +41,7 @@ void XYPlotter_delete(GI_uid_t uid) {
 	GI_delete(p_item);
 }
 
-GI_uid_t XYPlotter_drawPoint(int x, int y, Pen_t pen) {
+GI_uid_t XYPlotter_drawPoint(float x, float y, Pen_t pen) {
 	GraphicItem_t *p_item = GI_create_take();
 
 	p_item->sizeof_points = GO_drawPoint(NULL, x, y, pen.intensity);
@@ -53,7 +53,7 @@ GI_uid_t XYPlotter_drawPoint(int x, int y, Pen_t pen) {
 	return (GI_uid_t) p_item;
 }
 
-GI_uid_t XYPlotter_drawLine(int x1, int y1, int x2, int y2, Pen_t pen) {
+GI_uid_t XYPlotter_drawLine(float x1, float y1, float x2, float y2, Pen_t pen) {
 	GraphicItem_t *p_item = GI_create_take();
 
 	p_item->sizeof_points = GO_drawLine(NULL, x1, y1, x2, y2, pen.spacing, pen.intensity);
@@ -66,7 +66,7 @@ GI_uid_t XYPlotter_drawLine(int x1, int y1, int x2, int y2, Pen_t pen) {
 }
 
 
-GI_uid_t XYPlotter_drawRect(int x1, int y1, int x2, int y2, Pen_t pen){
+GI_uid_t XYPlotter_drawRect(float x1, float y1, float x2, float y2, Pen_t pen){
 	GraphicItem_t *p_item = GI_create_take();
 
 	p_item->sizeof_points = GO_drawRect(NULL, x1, y1, x2, y2, pen.spacing, pen.intensity);
@@ -79,7 +79,7 @@ GI_uid_t XYPlotter_drawRect(int x1, int y1, int x2, int y2, Pen_t pen){
 }
 
 
-GI_uid_t XYPlotter_drawArc(int x, int y, int r, float a1, float a2, Pen_t pen){
+GI_uid_t XYPlotter_drawArc(float x, float y, float r, float a1, float a2, Pen_t pen){
 	GraphicItem_t *p_item = GI_create_take();
 
 	p_item->sizeof_points = GO_drawArc(NULL, x, y, r, a1, a2, pen.spacing, pen.intensity);
@@ -92,7 +92,7 @@ GI_uid_t XYPlotter_drawArc(int x, int y, int r, float a1, float a2, Pen_t pen){
 }
 
 
-GI_uid_t XYPlotter_drawChar(char c, int x, int y, float height, Pen_t pen) {
+GI_uid_t XYPlotter_drawChar(char c, float x, float y, float height, Pen_t pen) {
 	GraphicItem_t *p_item = GI_create_take();
 
 	p_item->sizeof_points = GO_drawChar(NULL, c, x, y, height, pen.intensity);
@@ -104,7 +104,7 @@ GI_uid_t XYPlotter_drawChar(char c, int x, int y, float height, Pen_t pen) {
 	return (GI_uid_t) p_item;
 }
 
-GI_uid_t XYPlotter_drawString(const char* str, int x, int y, float height, Pen_t pen) {
+GI_uid_t XYPlotter_drawString(const char* str, float x, float y, float height, Pen_t pen) {
 	GraphicItem_t *p_item = GI_create_take();
 
 	p_item->sizeof_points = GO_drawString(NULL, str, x, y, height, pen.intensity);
